requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
boto3>=1.34.0
fastapi>=0.110.0
uvicorn>=0.29.0
ultralytics>=8.0.0
opencv-python>=4.8.0
duckduckgo-search>=6.0.0
mediapipe>=0.10.0
edge-tts>=7.0.0
# ── Phase B: Semantic memory ──────────────────────────────────────────────────
fastembed>=0.3.0
# ── Phase A: Whisper STT ──────────────────────────────────────────────────────
faster-whisper>=1.0.0
python-multipart>=0.0.9
# ── Phase A: Wake word ────────────────────────────────────────────────────────
openwakeword>=0.6.0
sounddevice>=0.4.6
# ── Phase A: Gmail ────────────────────────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
# ── Calendar (also needed for Gmail OAuth) ────────────────────────────────────
feedparser>=6.0.0
httpx>=0.27.0
yfinance>=0.2.0
# ── Phase C: MCP protocol ────────────────────────────────────────────────────
mcp>=1.0.0
# ── Phase C: Telegram bot ────────────────────────────────────────────────────
pyTelegramBotAPI>=4.14.0
# ── Phase C: Browser automation ──────────────────────────────────────────────
playwright>=1.40.0
# ── Phase D: System vitals ───────────────────────────────────────────────────
psutil>=5.9.0
# ── Phase C: Free dictation ──────────────────────────────────────────────────
pynput>=1.7.0
pyautogui>=0.9.54
pyperclip>=1.8.0
requests>=2.31.0
